pyproject.toml
[project]
name = "simple-backup"
version = "0.1.8"
description = "Простой инструмент для создания и восстановления бэкапов проектов в YAML формате"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pyyaml>=6.0",
    "click>=8.0",
]
authors = [
    {name = "vpuhoff", email = "vpuhoff92@gmail.com"}
]
keywords = ["backup", "yaml", "project", "restore", "cli"]
urls = {Homepage = "https://github.com/vpuhoff/simple-backup-tool", Repository = "https://github.com/vpuhoff/simple-backup-tool"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Archiving :: Backup",
    "Topic :: Utilities",
]

[project.scripts]
simple-backup = "main:cli"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "flake8>=6.0",
    "mypy>=1.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
